LayoutTests/platform/mac/accessibility/document-title-used-for-description.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<title>FirstTitle</title>
<script>
var successfullyParsed = false;
</script>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that <title> will be used as the description for a web area.");

    if (window.accessibilityController) {

        document.getElementById("body").focus();
        var webArea = window.accessibilityController.focusedElement;
        shouldBe("webArea.description", "'AXDescription: FirstTitle'");

        document.title = "SecondTitle";
        shouldBe("webArea.description", "'AXDescription: SecondTitle'");
    }

    successfullyParsed = true;
</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
