LayoutTests/platform/chromium/accessibility/document-title-used-for-description-expected.txt

This tests that

PASS webArea.description is 'AXDescription: FirstTitle'
PASS webArea.description is 'AXDescription: SecondTitle'
PASS successfullyParsed is true

TEST COMPLETE
